+++ examples/midi_player.rb
@@ -72,7 +72,7 @@ class TinyMidiPlayer
 
         combobox { |c|
           items @midi_files.map { |path| File.basename(path) }
-          # bind selected item (String)
+          # data-bind selected item (String) to self.selected_file with on-read/on-write converters and after_write operation
           selected_item <=> [self, :selected_file, on_read: ->(f) {File.basename(f.to_s)}, on_write: ->(f) {File.join(@music_directory, f)}, after_write: -> { play_midi if @th&.alive? }]
         }
       }